Move grammar tests to their own file
